requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
pyvi==0.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn-crfsuite==0.3.6
tabulate==0.8.7
threadpoolctl==2.1.0
tqdm==4.48.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
xgboost==1.2.0
